refactor(workersvc): move credential/secret selection to secretchoice.go
Pure same-package code motion (PRD #921 M2, seam 5a). Moves the secret-choice
cluster (autoLaneRetryable, staticChoice, openAnthropic, recordRunCredential,
claimSecretID, claimExclude, autoChoice, autoFloorRetry, workerSecretID) plus
the claimCred and secretChoice types and the selectReason* const block out of
service.go into the new secretchoice.go, byte-identically (autoChoice's inline
//nolint:gosec travels verbatim). ValidBindMode, the BindMode* consts, and the
shared claim error vars stay in service.go. The now-unused autoselectrow and
secretopen imports are dropped from service.go (pure deletions). No
exported-API change, no behavior change; go doc output unchanged.
